pyproject.toml
[project]
name = "tear-waste"
version = "0.1.0"
description = ""
authors = [
    {name = "egi",email = "egigjevori15@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"


dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
isort = "^6.0.1"
black = "^25.1.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
asyncpg-stubs = "^0.30.1"
flake8 = "^7.1.2"

[tool.pytest.ini_options]
asyncio_mode="auto"

[tool.mypy]
exclude = ["test/*", ".venv"]

[tool.flake8]
exclude = ".venv,test/*"
max-line-length = 120

[tool.black]
line-length = 120

[tool.isort]
py_version=39
line_length=120
profile="black"